refactor: remove unnecessary convert logic and move directory after renaming it
