+++ src/main/java/nextstep/subway/auth/ui/interceptor/authentication/SessionAuthenticationInterceptor.java
@@ -2,8 +2,6 @@ package nextstep.subway.auth.ui.interceptor.authentication;
 
 import static nextstep.subway.auth.infrastructure.SecurityContextHolder.*;
 
-import java.util.Map;
-
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
@@ -43,14 +41,6 @@ public class SessionAuthenticationInterceptor implements HandlerInterceptor {
         httpSession.setAttribute(SPRING_SECURITY_CONTEXT_KEY, new SecurityContext(authentication));
         response.setStatus(HttpServletResponse.SC_OK);
         return false;
-    }
-
-    public AuthenticationToken convert(HttpServletRequest request) {
-        Map<String, String[]> paramMap = request.getParameterMap();
-        String principal = paramMap.get(USERNAME_FIELD)[0];
-        String credentials = paramMap.get(PASSWORD_FIELD)[0];
-
-        return new AuthenticationToken(principal, credentials);
     }
 
     public Authentication authenticate(AuthenticationToken token) {
