+++ src/test/java/nextstep/subway/auth/ui/interceptor/authentication/SessionAuthenticationInterceptorTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.auth.ui.interceptor.jwt;
+package nextstep.subway.auth.ui.interceptor.authentication;
 
 import static nextstep.subway.auth.infrastructure.SecurityContextHolder.*;
 import static org.assertj.core.api.Assertions.*;
@@ -20,7 +20,6 @@ import org.springframework.mock.web.MockHttpServletResponse;
 import nextstep.subway.auth.domain.Authentication;
 import nextstep.subway.auth.domain.AuthenticationToken;
 import nextstep.subway.auth.infrastructure.SecurityContext;
-import nextstep.subway.auth.ui.interceptor.authentication.SessionAuthenticationInterceptor;
 import nextstep.subway.auth.ui.interceptor.convert.AuthenticationConverter;
 import nextstep.subway.member.application.CustomUserDetailsService;
 import nextstep.subway.member.domain.LoginMember;
